test: split prettier specific tests to their own file
