+++ integration-tests/cypress/e2e/test-utils.ts
@@ -2,6 +2,11 @@ import type { RunLuaCodeOutput } from "@tui-sandbox/library/server"
 import type { MyTestDirectoryFile } from "../../MyTestDirectory.js"
 import type { NeovimContext } from "../support/tui-sandbox.js"
 
+export const fakeGitRepoFileText =
+  "fake-git-repository-file-contents-71f64aabd056"
+
+export const lazygitWasClosedMessage = "tsugit.nvim: lazygit closed for"
+
 export function initializeGitRepositoryInDirectory(
   relativePath: MyTestDirectoryFile = "fakegitrepo",
 ): void {
