+++ integration-tests/test-environment/config-modifications/use_prettierd_for_formatting.lua
@@ -1,0 +1,7 @@
+require("tsugit").setup({
+  integrations = {
+    conform = {
+      formatter = "prettierd",
+    },
+  },
+})
